frameparams: adopt to API changes in zstd 1.3.1
The type of the window size member changed.

There are also new constants that the frameContentSize member
can be set to. Of note is that an unknown frame content size now
has value ZSTD_CONTENTSIZE_UNKNOWN instead of 0.

There is a new frame parsing API in 1.3.0. We should probably
switch to it at some point...

+++ c-ext/python-zstandard.h
@@ -49,7 +49,7 @@ extern PyTypeObject CompressionParametersType;
 typedef struct {
 	PyObject_HEAD
 	unsigned long long frameContentSize;
-	size_t windowSize;
+	unsigned long long windowSize;
 	unsigned dictID;
 	char checksumFlag;
 } FrameParametersObject;
